// utils/types/LargeProducersPublicList.types.ts
export type Nations = "England" | "Wales" | "Scotland" | "Northern Ireland";

export type ReportHeaders = {
  TitleEngland: string;
  TitleWales: string;
  TitleScotland: string;
  TitleNorthernIreland: string;
  TitleAllNations: string;
  Header1: string;
  Header2: string;
  Header3: string;
  Header4: string;
  Header5: string;
  Header6: string;
  Header7: string;
  Header8: string;
  Header9: string;
  Header10: string;
  Header11: string;
  Header12: string;
  Header13: string;
  Header14: string;
  Header15: string;
  Header16: string;
  Header17: string;
};